@@ -1,15 +1,12 @@
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Claims;
-using System.Threading.Tasks;
 using API.Data;
 using API.DTOs;
 using API.Entities;
 using API.Extensions;
+using API.Helpers;
 using API.Interfaces;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -21,7 +18,7 @@ namespace API.Controllers
         private readonly IUserRepository _userRepository;
         private readonly IMapper _mapper;
         private readonly IPhotoService _photoService;
-        public UsersController(IUserRepository userRepository, IMapper mapper,
+        public UsersController(IUserRepository userRepository, IMapper mapper, 
             IPhotoService photoService)
         {
             _photoService = photoService;
@@ -30,14 +27,26 @@ namespace API.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<MemberDto>>> GetUsers()
+        public async Task<ActionResult<PagedList<MemberDto>>> GetUsers([FromQuery]UserParams userParams)
         {
-            var users = await _userRepository.GetMembersAsync();
+            var currentUser = await _userRepository.GetUserByUsernameAsync(User.GetUsername());
+            userParams.CurrentUsername = currentUser.UserName;
+
+            if (string.IsNullOrEmpty(userParams.Gender))
+            {
+                userParams.Gender = currentUser.Gender == "male" ? "female" : "male";
+            }
+
+            var users = await _userRepository.GetMembersAsync(userParams);
+
+            Response.AddPaginationHeader(new PaginationHeader(users.CurrentPage, users.PageSize, 
+                users.TotalCount, users.TotalPages));
 
             return Ok(users);
         }
 
         [HttpGet("{username}")]
+
         public async Task<ActionResult<MemberDto>> GetUser(string username)
         {
             return await _userRepository.GetMemberAsync(username);
@@ -58,7 +67,6 @@ namespace API.Controllers
         }
 
         [HttpPost("add-photo")]
-
         public async Task<ActionResult<PhotoDto>> AddPhoto(IFormFile file)
         {
             var user = await _userRepository.GetUserByUsernameAsync(User.GetUsername());
@@ -66,6 +74,7 @@ namespace API.Controllers
             if (user == null) return NotFound();
 
             var result = await _photoService.AddPhotoAsync(file);
+
             if (result.Error != null) return BadRequest(result.Error.Message);
 
             var photo = new Photo
@@ -78,13 +87,13 @@ namespace API.Controllers
 
             user.Photos.Add(photo);
 
-            if (await _userRepository.SaveAllAsync())
+            if (await _userRepository.SaveAllAsync()) 
             {
                 return CreatedAtAction(nameof(GetUser), 
-                    new {username = user.UserName }, _mapper.Map<PhotoDto>(photo)) ;
+                    new {username = user.UserName}, _mapper.Map<PhotoDto>(photo));
             }
 
-            return BadRequest("Problem Adding photo");
+            return BadRequest("Problem adding photo");
         }
 
         [HttpPut("set-main-photo/{photoId}")]
@@ -120,7 +129,7 @@ namespace API.Controllers
 
             if (photo.IsMain) return BadRequest("You cannot delete your main photo");
 
-            if(photo.PublicId !=null)
+            if (photo.PublicId != null)
             {
                 var result = await _photoService.DeletePhotoAsync(photo.PublicId);
                 if (result.Error != null) return BadRequest(result.Error.Message);
