@@ -1,12 +1,9 @@
-using System;
-using System.Collections.Generic;
-
 namespace API.DTOs
 {
     public class MemberDto
     {
         public int Id { get; set; }
-        public string Username { get; set; }
+        public string UserName { get; set; }
         public string PhotoUrl { get; set; }
         public int Age { get; set; }
         public string KnownAs { get; set; }
@@ -18,6 +15,6 @@ namespace API.DTOs
         public string Interests { get; set; }
         public string City { get; set; }
         public string Country { get; set; }
-        public ICollection<PhotoDto> Photos { get; set; }
+        public List<PhotoDto> Photos { get; set; }
     }
 }
