@@ -1,5 +1,5 @@
-using System;
 using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Mvc.RazorPages.Infrastructure;
 
 namespace API.DTOs
 {
@@ -10,7 +10,7 @@ namespace API.DTOs
 
         [Required] public string KnownAs { get; set; }
         [Required] public string Gender { get; set; }
-        [Required] public DateTime? DateOfBirth { get; set; }
+        [Required] public DateOnly? DateOfBirth { get; set; } // optional to make required work!
         [Required] public string City { get; set; }
         [Required] public string Country { get; set; }
 
