@@ -1,9 +1,6 @@
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 using API.Entities;
 using Microsoft.EntityFrameworkCore;
 
@@ -15,18 +12,21 @@ namespace API.Data
         {
             if (await context.Users.AnyAsync()) return;
 
-            var userData = await System.IO.File.ReadAllTextAsync("Data/UserSeedData.json");
+            var userData = await File.ReadAllTextAsync("Data/UserSeedData.json");
+
+            var options = new JsonSerializerOptions{PropertyNameCaseInsensitive = true};
+
             var users = JsonSerializer.Deserialize<List<AppUser>>(userData);
-            if (users == null) return;
+
             foreach (var user in users)
             {
                 using var hmac = new HMACSHA512();
 
                 user.UserName = user.UserName.ToLower();
-                user.PasswordSalt = hmac.Key;
                 user.PasswordHash = hmac.ComputeHash(Encoding.UTF8.GetBytes("Pa$$w0rd"));
+                user.PasswordSalt = hmac.Key;
 
-                await context.Users.AddAsync(user);
+                context.Users.Add(user);
             }
 
             await context.SaveChangesAsync();
