@@ -1,5 +1,3 @@
-using System;
-using System.Collections.Generic;
 using API.Extensions;
 
 namespace API.Entities
@@ -10,10 +8,10 @@ namespace API.Entities
         public string UserName { get; set; }
         public byte[] PasswordHash { get; set; }
         public byte[] PasswordSalt { get; set; }
-        public DateTime DateOfBirth { get; set; }
+        public DateOnly DateOfBirth { get; set; }
         public string KnownAs { get; set; }
-        public DateTime Created { get; set; } = DateTime.Now;
-        public DateTime LastActive { get; set; } = DateTime.Now;
+        public DateTime Created { get; set; } = DateTime.UtcNow;
+        public DateTime LastActive { get; set; } = DateTime.UtcNow;
         public string Gender { get; set; }
         public string Introduction { get; set; }
         public string LookingFor { get; set; }
@@ -24,8 +22,7 @@ namespace API.Entities
 
         // public int GetAge()
         // {
-        //     return DateOfBirth.CalculateAge();
+        //     return DateOfBirth.CalcuateAge();
         // }
-
     }
 }
