@@ -9,7 +9,8 @@ namespace API.Entities
         public string Url { get; set; }
         public bool IsMain { get; set; }
         public string PublicId { get; set; }
-        public AppUser AppUser { get; set; }
+
         public int AppUserId { get; set; }
+        public AppUser AppUser { get; set; }
     }
 }
