@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace API.Extensions
 {
@@ -10,8 +6,12 @@ namespace API.Extensions
     {
         public static string GetUsername(this ClaimsPrincipal user)
         {
+            return user.FindFirst(ClaimTypes.Name)?.Value;
+        }
+
+        public static string GetUserId(this ClaimsPrincipal user)
+        {
             return user.FindFirst(ClaimTypes.NameIdentifier)?.Value;
         }
-        
     }
 }
