@@ -1,14 +1,15 @@
-using System;
-
 namespace API.Extensions
 {
     public static class DateTimeExtensions
     {
-        public static int CalculateAge(this DateTime dob)
+        public static int CalcuateAge(this DateOnly dob)
         {
-            var today = DateTime.Today;
+            var today = DateOnly.FromDateTime(DateTime.UtcNow);
+
             var age = today.Year - dob.Year;
-            if (dob.Date > today.AddYears(-age)) age--;
+
+            if (dob > today.AddYears(-age)) age--;
+
             return age;
         }
     }
