@@ -12,21 +12,24 @@ export class DatePickerComponent implements ControlValueAccessor {
   @Input() maxDate: Date | undefined;
   bsConfig: Partial<BsDatepickerConfig> | undefined;
 
-  constructor(@Self() public ngControl:NgControl) {
+  constructor(@Self() public ngControl: NgControl) {
     this.ngControl.valueAccessor = this;
-    this.bsConfig= {
+    this.bsConfig = {
       containerClass: 'theme-red',
-      dateInputFormat: 'MMM DD YYYY'
+      dateInputFormat: 'DD MMMM YYYY'
     }
    }
+
   writeValue(obj: any): void {
   }
+  
   registerOnChange(fn: any): void {
   }
+
   registerOnTouched(fn: any): void {
   }
 
-  get control(): FormControl{
+  get control(): FormControl {
     return this.ngControl.control as FormControl
   }
 
