@@ -8,11 +8,12 @@ import { ControlValueAccessor, FormControl, NgControl } from '@angular/forms';
 })
 export class TextInputComponent implements ControlValueAccessor {
   @Input() label = '';
-  @Input() type= 'text';
+  @Input() type = 'text';
 
   constructor(@Self() public ngControl: NgControl) { 
     this.ngControl.valueAccessor = this;
   }
+
   writeValue(obj: any): void {
   }
 
@@ -25,5 +26,4 @@ export class TextInputComponent implements ControlValueAccessor {
   get control(): FormControl {
     return this.ngControl.control as FormControl
   }
-
 }
