@@ -6,9 +6,8 @@ import { MemberEditComponent } from '../members/member-edit/member-edit.componen
   providedIn: 'root'
 })
 export class PreventUnsavedChangesGuard implements CanDeactivate<MemberEditComponent> {
-  canDeactivate(
-    component: MemberEditComponent) :  boolean {
-    if(component.editForm?.dirty){
+  canDeactivate(component: MemberEditComponent): boolean {
+    if (component.editForm?.dirty) {
       return confirm('Are you sure you want to continue? Any unsaved changes will be lost');
     }
     return true;
