@@ -5,9 +5,8 @@ import {
   HttpEvent,
   HttpInterceptor
 } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, take } from 'rxjs';
 import { AccountService } from '../_services/account.service';
-import { take } from 'rxjs/operators';
 
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
@@ -17,16 +16,16 @@ export class JwtInterceptor implements HttpInterceptor {
   intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
     this.accountService.currentUser$.pipe(take(1)).subscribe({
       next: user => {
-        if(user){
+        if (user) {
           request = request.clone({
             setHeaders: {
               Authorization: `Bearer ${user.token}`
             }
           })
-
         }
       }
-  })
+    })
+    
     return next.handle(request);
   }
 }
