@@ -15,12 +15,12 @@ export class LoadingInterceptor implements HttpInterceptor {
 
   intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
     this.busyService.busy();
-
+    
     return next.handle(request).pipe(
       delay(1000),
-      finalize (() => {
-        this.busyService.idle();
+      finalize(() => {
+        this.busyService.idle()
       })
-    );
+    )
   }
 }
