@@ -7,11 +7,12 @@ import { Router } from '@angular/router';
   styleUrls: ['./server-error.component.css']
 })
 export class ServerErrorComponent implements OnInit {
-error: any;
-  constructor(private router: Router) { 
+  error: any;
+
+  constructor(private router: Router) {
     const navigation = this.router.getCurrentNavigation();
-    this.error = navigation?.extras?.state?.error;
-  }
+    this.error = navigation?.extras?.state?.['error'];
+   }
 
   ngOnInit(): void {
   }
