@@ -6,7 +6,8 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
-registerMode = false;
+  registerMode = false;
+  users: any;
 
   constructor() { }
 
@@ -17,7 +18,8 @@ registerMode = false;
     this.registerMode = !this.registerMode;
   }
 
-  cancelRegisterMode(event: boolean){
-this.registerMode = event;
-  }
+  cancelRegisterMode(event: boolean) {
+    this.registerMode = event;
+  } 
+
 }
