@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { User } from '../_models/user';
 import { AccountService } from '../_services/account.service';
 
@@ -11,24 +11,23 @@ import { AccountService } from '../_services/account.service';
   styleUrls: ['./nav.component.css']
 })
 export class NavComponent implements OnInit {
-  model: any = {}
+  model: any = {};
+
   constructor(public accountService: AccountService, private router: Router, 
     private toastr: ToastrService) { }
 
   ngOnInit(): void {
-    
   }
 
-  login(){
-    this.accountService.login(this.model).subscribe(response => {
-      this.router.navigateByUrl('/members')
-      
+  login() {
+    this.accountService.login(this.model).subscribe({
+      next: _ => this.router.navigateByUrl('/members')
     })
   }
 
   logout() {
     this.accountService.logout();
-    this.router.navigateByUrl('/')
+    this.router.navigateByUrl('/');
   }
 
 }
