@@ -15,12 +15,12 @@ export class RegisterComponent implements OnInit {
   maxDate: Date = new Date();
   validationErrors: string[] | undefined;
 
-  constructor(private accountService: AccountService, private toastr: ToastrService,
-    private fb: FormBuilder, private router: Router) { }
+  constructor(private accountService: AccountService, private toastr: ToastrService, 
+      private fb: FormBuilder, private router: Router) { }
 
   ngOnInit(): void {
     this.initializeForm();
-    this.maxDate.setFullYear(this.maxDate.getFullYear() - 18);
+    this.maxDate.setFullYear(this.maxDate.getFullYear() -18);
   }
 
   initializeForm() {
@@ -31,8 +31,8 @@ export class RegisterComponent implements OnInit {
       dateOfBirth: ['', Validators.required],
       city: ['', Validators.required],
       country: ['', Validators.required],
-      password: ['', [Validators.required,
-      Validators.minLength(4), Validators.maxLength(8)]],
+      password: ['', [Validators.required, 
+        Validators.minLength(4), Validators.maxLength(8)]],
       confirmPassword: ['', [Validators.required, this.matchValues('password')]],
     });
     this.registerForm.controls['password'].valueChanges.subscribe({
@@ -42,7 +42,7 @@ export class RegisterComponent implements OnInit {
 
   matchValues(matchTo: string): ValidatorFn {
     return (control: AbstractControl) => {
-      return control.value === control.parent?.get(matchTo)?.value ? null : { notmatching: true }
+      return control.value === control.parent?.get(matchTo)?.value ? null : {notMatching: true}
     }
   }
 
@@ -55,16 +55,19 @@ export class RegisterComponent implements OnInit {
       },
       error: error => {
         this.validationErrors = error
-      }
+      } 
     })
   }
+
   cancel() {
     this.cancelRegister.emit(false);
   }
 
-  private getDateOnly(dob: string | undefined){
+  private getDateOnly(dob: string | undefined) {
     if (!dob) return;
     let theDob = new Date(dob);
-    return new Date(theDob.setMinutes(theDob.getMinutes()- theDob.getTimezoneOffset())).toISOString().slice(0,10);
+    return new Date(theDob.setMinutes(theDob.getMinutes()-theDob.getTimezoneOffset()))
+      .toISOString().slice(0,10);
   }
+
 }
